+++ collector/src/main/scala/org/opentorah/collector/ByHierarchy.scala
@@ -1,7 +1,6 @@
 package org.opentorah.collector
 
-import org.opentorah.site.HtmlContent
-import org.opentorah.store.{By, Path, Pure}
+import org.opentorah.store.{By, Context, Path, Pure}
 import org.opentorah.xml.{Element, Parsable, Parser, ScalaXml, Unparser}
 
 final class ByHierarchy(
@@ -18,19 +17,19 @@ final class ByHierarchy(
   // TODO generate hierarchy root index and reference it from the summary.
   // TODO allow viewing tree indexes rooted in intermediate ByHierarchys.
 
-  def oneLevelIndex(path: Path, collector: Collector): ScalaXml.Element =
+  def oneLevelIndex(path: Path, pathShortener: Path.Shortener): ScalaXml.Element =
     <p>
       <l>{Hierarchical.displayName(this)}:</l>
       <ul>{
         for hierarchical <- hierarchyStores yield
           val hierarchicalPath: Path = path :+ hierarchical
           <li>
-            {reference(hierarchicalPath, hierarchical, collector)}
+            {hierarchical.reference(hierarchicalPath, pathShortener)}
           </li>
       }</ul>
     </p>
 
-  def treeIndex(path: Path, collector: Collector): ScalaXml.Element =
+  def treeIndex(path: Path, context: Context, pathShortener: Path.Shortener): ScalaXml.Element =
     <div class="tree-index">
       <ul>
         <li><em>{Hierarchical.displayName(this)}</em></li>
@@ -39,16 +38,13 @@ final class ByHierarchy(
             for hierarchical <- hierarchyStores yield
               val hierarchicalPath: Path = path :+ hierarchical
               <li>
-                {reference(hierarchicalPath, hierarchical, collector)}
-                {hierarchical.getBy.toSeq.map(by => by.treeIndex(hierarchicalPath :+ by, collector))}
+                {hierarchical.reference(hierarchicalPath, pathShortener)}
+                {hierarchical.getBy.toSeq.map(by => by.treeIndex(hierarchicalPath :+ by, context, pathShortener))}
               </li>
           }</ul>
         </li>
       </ul>
     </div>
-
-  private def reference(path: Path, hierarchical: Hierarchical, collector: Collector): ScalaXml.Element =
-    collector.a(path)(text = hierarchical.displayTitle)
 
 object ByHierarchy extends Element[ByHierarchy]("by"):
   override def contentParsable: Parsable[ByHierarchy] = new Parsable[ByHierarchy]: