+++ collector/src/main/scala/org/opentorah/collector/EntityList.scala
@@ -1,8 +1,7 @@
 package org.opentorah.collector
 
 import org.opentorah.metadata.Names
-import org.opentorah.site.HtmlContent
-import org.opentorah.store.{Path, Pure}
+import org.opentorah.store.{Context, Path, Pure, Viewer}
 import org.opentorah.tei.{EntityRelated, EntityType, Title}
 import org.opentorah.xml.{Attribute, Caching, Element, Parsable, Parser, ScalaXml, Unparser}
 import zio.ZIO
@@ -14,7 +13,10 @@ final class EntityList(
   val entityType: EntityType,
   val role: Option[String],
   val title: Title.Value,
-) extends Pure[Entity], HtmlContent.ApparatusViewer[Collector], Element.FromUrl.With:
+) extends
+  Pure[Entity],
+  Viewer.Apparatus,
+  Element.FromUrl.With:
   private var entities: Seq[Entity] = Seq.empty
 
   def setEntities(value: Seq[Entity]): Unit =
@@ -27,11 +29,11 @@ final class EntityList(
   override def htmlHeadTitle: Option[String] = Some(title.content.toString)
   override def htmlBodyTitle: Option[ScalaXml.Nodes] = Some(title.content.scalaXml)
 
-  override def content(path: Path, collector: Collector): Caching.Parser[ScalaXml.Element] = ZIO.succeed(
+  override def content(path: Path, context: Context): Caching.Parser[ScalaXml.Element] =
+    for pathShortener: Path.Shortener <- context.pathShortener yield
     <list>
-      {for entity <- getEntities yield Entity.line(entity, collector)}
+      {for entity <- getEntities yield entity.line(context, pathShortener)}
     </list>
-  )
 
 object EntityList extends EntityRelated[EntityList](
   elementName = _.listElement,
@@ -39,9 +41,9 @@ object EntityList extends EntityRelated[EntityList](
 ):
   override protected def contentType: Element.ContentType = Element.ContentType.Elements
 
-  private val roleAttribute: Attribute.Optional[String] = Attribute("role").optional
-
   override protected def parsable(entityType: EntityType): Parsable[EntityList] = new Parsable[EntityList]:
+    private val roleAttribute: Attribute.Optional[String] = Attribute("role").optional
+
     override def parser: Parser[EntityList] = for
       fromUrl: Element.FromUrl <- Element.fromUrl
       names: Names <- Names.withDefaultNameParsable()