+++ site/src/main/scala/org/opentorah/store/Context.scala
@@ -1,0 +1,9 @@
+package org.opentorah.store
+
+import org.opentorah.xml.Caching
+
+trait Context {
+  def pathShortener: Caching.Parser[Path.Shortener]
+  
+  def path(store: Store): Path
+}